users can now reset password

+++ app/api/comparetoken/route.ts
@@ -11,10 +11,19 @@ export async function POST(req: Request) {
   if (!currentToken){
     return NextResponse.json({error:"Invalid token"})
   }
+  if(parseInt(currentToken.expires) < Date.now()){
+    return NextResponse.json({error:"Token expired"})
+  }
   const user = await users.findOne({_id:currentToken.userId})
   if(user){
+    const update ={
+      $set:{
+        "passChange":true
+      }
+    }
+     users.updateOne({_id:user._id}, update)
      tokens.deleteOne({token})
      return NextResponse.json(user)
   }
-  return NextResponse.json({erro:"User Not found"})
+  return NextResponse.json({error:"User Not found"})
 }
